+++ src/lib.rs
@@ -1,19 +1,18 @@
+mod helpers;
+
+use helpers::*;
 use proc_macro::TokenStream;
 use quote::quote;
-use syn::{parse_macro_input, DeriveInput, Fields, Ident, Type};
+use syn::{parse_macro_input, DeriveInput, Ident, Type};
 
-#[proc_macro_derive(AutoGet)]
-pub fn auto_get_derive(input: TokenStream) -> TokenStream {
+#[proc_macro_derive(AutoGetters)]
+pub fn auto_getters(input: TokenStream) -> TokenStream {
     let input: DeriveInput = parse_macro_input!(input as DeriveInput);
     let name: Ident = input.ident;
 
-    let fields: Fields = if let syn::Data::Struct(data) = input.data {
-        data.fields
-    } else {
-        return TokenStream::new()
-    };
+    let fields = get_fields(&input.data);
 
-    let getters = fields.iter().map(|f| {
+    let getters_methods = fields.iter().map(|f| {
         let f_name: &Ident = f.ident.as_ref().unwrap();
         let f_type: &Type = &f.ty;
 
@@ -24,11 +23,52 @@ pub fn auto_get_derive(input: TokenStream) -> TokenStream {
         }
     });
 
-    let expanded: TokenStream = quote! {
+    let output: TokenStream = quote! {
         impl #name {
-            #(#getters)*
+            #(#getters_methods)*
         }
     }.into();
 
-    expanded
+    output
+}
+
+#[proc_macro_attribute]
+pub fn optional(_attr: TokenStream, item: TokenStream) -> TokenStream {
+    let input: DeriveInput = parse_macro_input!(item as DeriveInput);
+    let name: Ident = input.ident;
+
+    let fields = get_fields(&input.data);
+
+    let optional_fields = fields.iter().map(|f| {
+        let f_name = f.ident.as_ref().unwrap();
+        let f_type = &f.ty;
+
+        let mut is_except = false;
+
+        for attr in &f.attrs {
+            let path = attr.path();
+
+            if path.is_ident("except") {
+                is_except = true;
+            }
+        }
+
+        if is_except {
+            quote! {
+                #f_name: #f_type
+            }
+        } else {
+            quote! {
+                #f_name: Option<#f_type>
+            }
+        }
+    });
+
+    let output: TokenStream = quote! {
+        struct #name {
+            #(#optional_fields),*
+        }
+    }.into();
+
+    output
 }