+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## [1.1.0]
+
+- Add `optional` procedural macro with attribute for structs
+
 ## [1.0.1]
 
 ### Fixed
@@ -10,4 +14,4 @@
 
 ### Added
 
-- Add `auto_getters()` procedural macro for struct
+- Add `auto_getters()` procedural macro for structs